🐛 Fix: 타입 분리 (#7)
Fix: 타입 분리 (#7)
