+++ src/types/DiscussionPage/discussion.ts
@@ -1,0 +1,9 @@
+import type { Book } from '@/types/HomePage/home';
+
+export type BookCardProps = {
+  book: Book;
+};
+
+export type BookListProps = {
+  books: Book[];
+};
